+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 5.5.0beta2 - 2025-01-30
+
 ### Changed
 
 - Clarified what `PangeaConfig.environment` affects.